Week13/MiniProjet2/requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0

# LLM backends
groq>=0.4.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Machine Learning (for insights server)
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0.1
loguru>=0.7.0
tenacity>=8.2.0
aiofiles>=23.2.1

# Development
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
mypy>=1.5.0
